consolidate AssetIcon component
