+++ apps/webapp/src/components/swap/asset-out-selector.tsx
@@ -1,6 +1,6 @@
 import { AssetBalance } from '../../fetchers/balances';
 import { Dialog, DialogClose, DialogContent, DialogHeader, DialogTrigger } from '@penumbra-zone/ui';
-import { AssetIcon } from '../shared/asset-icon';
+import { AssetIcon } from '@penumbra-zone/ui/components/ui/tx/view/asset-icon';
 import { Metadata } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
 import { localAssets } from '@penumbra-zone/constants';
